+++ CHANGELOG.md
@@ -19,7 +19,7 @@ at anytime.
 
 ### Fixed
   * Race condition from improper initialization and shutdown of the blob manager database
-  *
+  * Various fixes for GetStream class used in API command get
 
 ### Deprecated
   *